CSI-220: creating modules out of execuer + device connectivity
