+++ node/pkg/driver/device_connectivity/device_connectivity_iscsi.go
@@ -1,3 +1,19 @@
+/**
+ * Copyright 2019 IBM Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package device_connectivity
 
 import (
@@ -9,7 +25,7 @@ import (
 	"time"
     "sync"
 	"path/filepath"
- 	"github.com/ibm/ibm-block-csi-driver/node/driver/executer"
+ 	executer "github.com/ibm/ibm-block-csi-driver/node/pkg/driver/executer"
 )
 
 type OsDeviceConnectivityIscsi struct {
@@ -19,7 +35,7 @@ type OsDeviceConnectivityIscsi struct {
 }
 
 
-func NewOsDeviceConnectivityIscsi(executer ExecuterInterface) OsDeviceConnectivityInterface{
+func NewOsDeviceConnectivityIscsi(executer executer.ExecuterInterface) OsDeviceConnectivityInterface{
 	return &OsDeviceConnectivityIscsi{
 		executer: executer, 
 		mutexMultipathF: &sync.Mutex{}, 
@@ -285,11 +301,11 @@ type OsDeviceConnectivityHelperIscsiInterface interface {
 }
 
 type OsDeviceConnectivityHelperIscsi struct {
-	executer  ExecuterInterface
+	executer  executer.ExecuterInterface
 }
 
 
-func NewOsDeviceConnectivityHelperIscsi(executer ExecuterInterface) OsDeviceConnectivityHelperIscsiInterface{
+func NewOsDeviceConnectivityHelperIscsi(executer executer.ExecuterInterface) OsDeviceConnectivityHelperIscsiInterface{
 	return &OsDeviceConnectivityHelperIscsi{executer: executer}
 }
