+++ node/pkg/driver/device_connectivity/messages.go
@@ -1,0 +1,4 @@
+package device_connectivity
+
+
+var ErrorNothingWasWrittenToScanFile = "Nothing was written to file : {%v}"
