+++ node/pkg/driver/errors.go
@@ -36,17 +36,6 @@ func (e *RequestValidationError) Error() string {
 	return fmt.Sprintf("Request Validation Error: %s", e.Msg)
 }
 
-
-type ConnectivityIscsiStorageTargetNotFoundError struct {
-	StorageTargetName string
-	DirectoryPath string
-}
-
-func (e *ConnectivityIscsiStorageTargetNotFoundError) Error() string {
-	return fmt.Sprintf("Connectivity Error: Storage target name [%s] was not found on the host, under directory %s", e.StorageTargetName, e.DirectoryPath)
-}
-
-
 type VolumeAlreadyProcessingError struct {
 	volId string
 }
@@ -54,49 +43,3 @@ type VolumeAlreadyProcessingError struct {
 func (e *VolumeAlreadyProcessingError) Error() string {
 	return fmt.Sprintf("Volume %s is already processing. request cannot be completed.", e.volId)
 }
-
-
-
-type MultipleDmDevicesError struct {
-	VolumeId string
-	LunId int
-	ArrayIqn string
-	MultipathDevicesMap map[string]bool
-}
-
-func (e *MultipleDmDevicesError) Error() string {
-	var mps string
-	for key:=range e.MultipathDevicesMap{
-		mps += ", " + key
-	}
-	return fmt.Sprintf("Detected more then one multipath devices (%s) for single volume (%s) with lunID %d from array target iqn %s",mps , e.VolumeId, e.LunId, e.ArrayIqn)
-}
-
-
-
-
-type MultipleDeviceNotFoundError struct {
-	DiskByPathDevice string
-	LinkToPhysicalDevice string
-}
-
-func (e *MultipleDeviceNotFoundError) Error() string {
-	return fmt.Sprintf("Couldn't find dm-* of the physical device path [%s -> %s] ", e.DiskByPathDevice, e.LinkToPhysicalDevice)
-}
-
-
-
-
-type MultipleDeviceNotFoundForLunError struct {
-	VolumeId string
-	LunId int
-	ArrayIqn string
-}
-
-func (e *MultipleDeviceNotFoundForLunError) Error() string {
-	return fmt.Sprintf("Couldn't find multipath device for volumeID [%s] lunID [%d] from array [%s]", e.VolumeId, e.LunId, e.ArrayIqn)
-}
-
-
-
-
