+++ node/pkg/driver/messages.go
@@ -3,5 +3,3 @@ package driver
 var ErrorWhileTryingToReadIQN = "Error while trying to get iqn  from string: %v."
 
 var ErrorUnsupportedConnectivityType = "Unsupported connectivity type : {%v}"
-
-var ErrorNothingWasWrittenToScanFile = "Nothing was written to file : {%v}"
